refactor requested
